Add alternate keywords for delete
The command only uses delete as the command keyword to execute the
command.

Let's add more alternate keywords del, rm, remove which can also be used
in place of delete.

+++ src/main/java/scrolls/elder/logic/commands/DeleteCommand.java
@@ -12,18 +12,30 @@ import scrolls.elder.model.Model;
 import scrolls.elder.model.person.Person;
 
 /**
- * Deletes a person identified using it's displayed index from the address book.
+ * Deletes a person identified using its displayed index from the address book.
  */
 public class DeleteCommand extends Command {
 
-    public static final String COMMAND_WORD = "delete";
+    public static final String COMMAND_WORD_DElETE = "delete";
+    public static final String COMMAND_WORD_DEL = "del";
+    public static final String COMMAND_WORD_RM = "rm";
+    public static final String COMMAND_WORD_REMOVE = "remove";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD
+    public static final String MESSAGE_USAGE = COMMAND_WORD_DElETE
             + ": Deletes the person identified by the index number used in the displayed person list.\n"
             + "Parameters: INDEX (must be a positive integer)\n"
-            + "Example: " + COMMAND_WORD + " 1";
+            + "Example: " + COMMAND_WORD_DElETE + " 1"
+            + "Alternatively, you can also delete a person using the following commands as well.\n"
+            + "Example: " + COMMAND_WORD_DEL + " 1"
+            + "Example: " + COMMAND_WORD_RM + " 1"
+            + "Example: " + COMMAND_WORD_REMOVE + " 1";
+
+
 
     public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
+    public static final String MESSAGE_DELETE_PERSON_ERROR = "Unable to delete contact: ";
+    public static final String MESSAGE_CONFIRM_DELETE =
+            "Valid contact inputted. Are you sure you want to delete this contact?";
 
     private final Index targetIndex;
 
@@ -37,10 +49,18 @@ public class DeleteCommand extends Command {
         List<Person> lastShownList = model.getFilteredPersonList();
 
         if (targetIndex.getZeroBased() >= lastShownList.size()) {
-            throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
+            throw new CommandException(MESSAGE_DELETE_PERSON_ERROR + Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
         }
 
         Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        /**
+         * Check if the person to be deleted is paired with another person.
+         */
+        if (personToDelete.isPairPresent(personToDelete)) {
+            throw new CommandException(MESSAGE_DELETE_PERSON_ERROR + Messages.MESSAGE_CONTACT_PAIRED_BEFORE_DELETE);
+        }
+
         model.deletePerson(personToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
     }
